restructures repository
