+++ Pipfile
@@ -1,0 +1,20 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+networkx = "==2.6.3"
+pandas = "==1.3.4"
+pyshp = "==2.1.3"
+tk = "==0.1.0"
+# Dependencies from VCS through ssh
+# pipenv install -e git+ssh://git@gitlab.com/konnecta/panthora-eye/modules/logging-service.git@v2.0.1#egg=logsrv
+python-dotenv = "*"
+# logsrv = {editable = true, ref = "v2.0.1", git = "ssh://git@gitlab.com/konnecta/panthora-eye/modules/logging-service.git"}
+
+[dev-packages]
+pylint = "*"
+
+[requires]
+python_version = "3.8.6"
