+++ setup.py
@@ -1,0 +1,72 @@
+"""A setuptools based setup module for the parcel generation.
+"""
+
+import pathlib
+from os.path import join
+from ast import literal_eval
+
+from io import open
+
+# Always prefer setuptools over distutils
+from setuptools import setup, find_packages
+
+
+here = pathlib.Path(__file__).parent.resolve()
+long_description = (here / 'README.md').read_text(encoding='utf-8')
+
+# Get the version from the __init__.py file
+def get_version():
+    """Scan __init__ file for __version__ and retrieve."""
+
+    finit = join(here, 'src', 'parcelgen', '__init__.py')
+    with open(finit, 'r', encoding="utf-8") as fpnt:
+        for line in fpnt:
+            if line.startswith('__version__'):
+                return literal_eval(line.split('=', 1)[1].strip())
+    return '0.0.1-alpha'
+
+setup(
+    name='parcel-gen',
+    version=get_version(),
+    description='A sample Python project',
+    long_description=long_description,
+    long_description_content_type='text/markdown',
+    url='https://github.com/Horizon-LEAD/Parcel_Generation',
+    keywords='lead, parcel, generation',
+    package_dir={'': 'src'},
+    packages=find_packages(where='src'),
+    python_requires='>=3.6, <4',
+    install_requires=[
+        'networkx',
+        'pandas',
+        'pyshp'
+    ],
+
+    # extras_require={  # Optional
+    #     'dev': ['check-manifest'],
+    #     'test': ['coverage'],
+    # },
+    
+    # # If there are data files included in your packages that need to be
+    # # installed, specify them here.
+    # package_data={  # Optional
+    #     'sample': ['package_data.dat'],
+    # },
+
+    # # Although 'package_data' is the preferred approach, in some case you may
+    # # need to place data files outside of your packages. See:
+    # # http://docs.python.org/distutils/setupscript.html#installing-additional-files
+    # #
+    # # In this case, 'data_file' will be installed into '<sys.prefix>/my_data'
+    # data_files=[('my_data', ['data/data_file'])],  # Optional
+    
+    entry_points={
+        'console_scripts': [
+            'parcelgen=parcelgen.__main__:main'
+        ],
+    },
+    project_urls={
+        'Bug Reports': 'https://github.com/Horizon-LEAD/Parcel_Generation/issues',
+        'Source': 'https://github.com/Horizon-LEAD/Parcel_Generation',
+    }
+)
