+++ src/parcelgen/__init__.py
@@ -1,0 +1,2 @@
+__version__="0.1.0"
+__version_info__=(0, 1, 0)
