+++ src/parcelgen/ui.py
@@ -1,0 +1,71 @@
+# import zlib
+# import base64
+# import tempfile
+from threading import Thread
+from pdb import set_trace
+
+import tkinter as tk
+from tkinter.ttk import Progressbar
+
+from .proc import run_model
+
+
+class ParcelGenUI:    
+    def __init__(self, args):       
+        '''
+        Initialize a GUI object
+        '''        
+        # Set graphics parameters
+        self.width  = 500
+        self.height = 60
+        self.bg     = 'black'
+        self.fg     = 'white'
+        self.font   = 'Verdana'
+        
+        # Create a GUI window
+        self.root = tk.Tk()
+        self.root.title("Progress Parcel Demand")
+        self.root.geometry(f'{self.width}x{self.height}+0+200')
+        self.root.resizable(False, False)
+        self.canvas = tk.Canvas(self.root, width=self.width, height=self.height, bg=self.bg)
+        self.canvas.place(x=0, y=0)
+        self.statusBar = tk.Label(self.root, text="", anchor='w', borderwidth=0, fg='black')
+        self.statusBar.place(x=2, y=self.height-22, width=self.width, height=22)
+        
+        # Remove the default tkinter icon from the window
+        # icon = zlib.decompress(base64.b64decode('eJxjYGAEQgEBBiDJwZDBy''sAgxsDAoAHEQCEGBQaIOAg4sDIgACMUj4JRMApGwQgF/ykEAFXxQRc='))
+        # _, self.iconPath = tempfile.mkstemp()
+        # with open(self.iconPath, 'wb') as iconFile:
+        #     iconFile.write(icon)
+        # self.root.iconbitmap(bitmap=self.iconPath)
+        
+        # Create a progress bar
+        self.progressBar = Progressbar(self.root, length=self.width-20)
+        self.progressBar.place(x=10, y=10)
+        
+        self.returnInfo = ""
+        self.args = args
+        set_trace()
+
+        self.run_module()       
+        
+        # Keep GUI active until closed    
+        self.root.mainloop()
+
+    def update_statusbar(self, text):
+        self.statusBar.configure(text=text)
+
+    def error_screen(self, text='', event=None, size=[800,50], title='Error message'):
+        '''
+        Pop up a window with an error message
+        '''
+        windowError = tk.Toplevel(self.root)
+        windowError.title(title)
+        windowError.geometry(f'{size[0]}x{size[1]}+0+{200+50+self.height}')
+        windowError.minsize(width=size[0], height=size[1])
+        windowError.iconbitmap(default=self.iconPath)
+        labelError = tk.Label(windowError, text=text, anchor='w', justify='left')
+        labelError.place(x=10, y=10)  
+
+    def run_module(self, event=None):
+        Thread(target=run_model, args=(self.args,), kwargs={'root': self}, daemon=True).start()
